Mark SingleThreadAsserter as @LegacyArchitecture
Summary:
Mark SingleThreadAsserter as LegacyArchitecture

changelog: [internal] internal

Differential Revision: D71820293
